requirements.txt
streamlit
pymupdf
chromadb
sentence-transformers
Pylance
openai
faiss-cpu
jsonlib